requirements.txt
ultralytics
opencv-python-headless
matplotlib
python-dotenv
roboflow
iPython